glasses/requirements_glasses.txt
flask>=3.0
adafruit-circuitpython-dht
picamera2
luma.oled
Pillow
